Support for pawn promotions
Added MoveResult enum, which can either be Normal, Promotion or Illegal.

If it is Promotion, the move is not executed immediately, but instead, stored as promotion_move in the board state. This must then be resolved before another move can be made. It can be resolved in two ways. Either, we can promote the pawn with Board::resolve_promotion() and provide a PieceType to promote to, in which case the move gets executed.

Or we can cancel the promotion with Board::cancel_promotion() in which case no move is executed